python/infobip-calls-mediastream-client/requirements.txt
numpy==1.24.3
PyYAML==6.0
scipy==1.10.1
websockets==11.0.3